A Basic Survival game for OOP repetetion
While designing this game, I paid attention to the implementation of many classes and OOP rules. And in this way, I reinforced the object-oriented programming OOP features (abstraction, encapsulation, inheritance and polymorphism) that are found in most programming languages ​​and are very important for coding, and used them in a simple project.